+++ packages/qb/pqb/src/sql/data.ts
@@ -50,6 +50,7 @@ export type CommonQueryData = {
   joinTo?: QueryBase;
   joinedShapes?: JoinedShapes;
   joinedParsers?: JoinedParsers;
+  joinRelationAs?: string;
   schema?: string;
   select?: SelectItem[];
   as?: string;